# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CRM Comercial B2B API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /emails:
    get:
      operationId: getEmails
      tags: [emails]
      summary: List emails
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: assignedTo
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of emails
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailListResponse"
  /emails/{id}:
    get:
      operationId: getEmail
      tags: [emails]
      summary: Get email
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Email"
  /emails/{id}/classify:
    post:
      operationId: classifyEmail
      tags: [emails]
      summary: Manually classify an email
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClassifyEmailRequest"
      responses:
        "200":
          description: Email classified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Email"
  /emails/{id}/process:
    post:
      operationId: processEmail
      tags: [emails]
      summary: Process email with AI (extract data, create opportunity)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Email processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailProcessResult"
  /emails/{id}/reply-draft:
    post:
      operationId: generateReplyDraft
      tags: [emails]
      summary: Generate AI reply draft for email
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Draft generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReplyDraft"
  /emails/{id}/send-reply:
    post:
      operationId: sendEmailReply
      tags: [emails]
      summary: Send reply email via Gmail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendReplyRequest"
      responses:
        "200":
          description: Reply sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /gmail/callback:
    get:
      operationId: gmailCallback
      tags: [gmail]
      summary: Gmail OAuth callback
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GmailConnection"
...
  /imports/clients:
    post:
      operationId: importClients
      tags: [imports]
      summary: Import clients from CSV
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CsvImportRequest"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
  /imports/products:
    post:
      operationId: importProducts
      tags: [imports]
      summary: Import products/measurements from CSV
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CsvImportRequest"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
  /imports/contacts:
    post:
      operationId: importContacts
      tags: [imports]
      summary: Import contacts from CSV
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CsvImportRequest"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Product:
      type: object
      properties:
        id:
          type: integer
        code:
          type: ["string", "null"]
        name:
          type: string
        description:
          type: ["string", "null"]
        unit:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        dimensions:
          type: ["string", "null"]
        standard:
          type: ["string", "null"]
        price:
          type: ["number", "null"]
        currency:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - isActive
        - createdAt
...
    Email:
      type: object
      properties:
        id:
          type: integer
        gmailMessageId:
          type: ["string", "null"]
        gmailThreadId:
          type: ["string", "null"]
        subject:
          type: string
        fromEmail:
          type: string
        fromName:
          type: ["string", "null"]
        toEmail:
          type: string
        body:
          type: string
        bodyHtml:
          type: ["string", "null"]
        receivedAt:
          type: string
          format: date-time
        category:
          type: ["string", "null"]
          enum: [quote_request, complaint, inquiry, follow_up, supplier, internal, spam, other, null]
        categoryConfidence:
          type: ["number", "null"]
        categoryConfirmed:
          type: boolean
        status:
          type: string
          enum: [pending, processing, processed, replied, closed]
        assignedSalespersonId:
          type: ["integer", "null"]
        clientId:
          type: ["integer", "null"]
        opportunityId:
          type: ["integer", "null"]
        extractedData:
          type: ["object", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - subject
        - fromEmail
        - toEmail
        - body
        - receivedAt
        - categoryConfirmed
        - status
        - createdAt
    EmailListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Email"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit
    ClassifyEmailRequest:
      type: object
      properties:
        category:
          type: string
          enum: [quote_request, complaint, inquiry, follow_up, supplier, internal, spam, other]
        notes:
          type: string
      required:
        - category
    EmailProcessResult:
      type: object
      properties:
        email:
          $ref: "#/components/schemas/Email"
        opportunity:
          $ref: "#/components/schemas/Opportunity"
        extractedData:
          type: object
        matchedProducts:
          type: array
          items:
            $ref: "#/components/schemas/Product"
      required:
        - email
        - extractedData
    ReplyDraft:
      type: object
      properties:
        subject:
          type: string
        body:
          type: string
        bodyHtml:
          type: string
      required:
        - subject
        - body
    SendReplyRequest:
      type: object
      properties:
        subject:
          type: string
        body:
          type: string
      required:
        - subject
        - body
...
    GmailConnection:
      type: object
      properties:
        connected:
          type: boolean
        email:
          type: ["string", "null"]
        lastSyncAt:
          type: ["string", "null"]
          format: date-time
      required:
        - connected
...
    Opportunity:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        clientId:
          type: ["integer", "null"]
        contactId:
          type: ["integer", "null"]
        salespersonId:
          type: ["integer", "null"]
        emailId:
          type: ["integer", "null"]
        status:
          type: string
          enum: [new, quoted, negotiating, won, lost, closed]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        estimatedValue:
          type: ["number", "null"]
        currency:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        extractedProducts:
          type: ["object", "null"]
        nextFollowUpAt:
          type: ["string", "null"]
          format: date-time
        closedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - status
        - priority
        - createdAt
        - updatedAt
...
    CsvImportRequest:
      type: object
      properties:
        data:
          type: array
          items:
            type: object
        mode:
          type: string
          enum: [insert, upsert]
      required:
        - data
        - mode
    ImportResult:
      type: object
      properties:
        total:
          type: integer
        inserted:
          type: integer
        updated:
          type: integer
        errors:
          type: integer
        errorDetails:
          type: array
          items:
            type: object
      required:
        - total
        - inserted
        - updated
        - errors
        - errorDetails
...